requirements.txt
aiogram==3.13.1
httpx==0.27.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
reportlab==4.2.5
python-dotenv==1.0.1
aiohttp==3.10.10
libsql==0.4.0
sqlalchemy-libsql==0.1.0
